+++ src/main/java/org/codeit/roomunion/common/exception/BaseErrorCode.java
@@ -1,0 +1,12 @@
+package org.codeit.roomunion.common.exception;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseErrorCode {
+
+    int getCode();
+
+    String getMessage();
+
+    HttpStatus getStatus();
+}
