--- src/main/java/org/codeit/roomunion/common/exception/model/BaseErrorCode.java
@@ -1,12 +1,0 @@
-package org.codeit.roomunion.common.exception.model;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseErrorCode {
-
-    int getCode();
-
-    String getMessage();
-
-    HttpStatus getStatus();
-}
