--- src/main/java/org/codeit/roomunion/crew/application/port/in/CrewCommandUseCase.java
@@ -1,9 +1,0 @@
-package org.codeit.roomunion.crew.application.port.in;
-
-import org.codeit.roomunion.crew.domain.model.Crew;
-import org.codeit.roomunion.crew.domain.model.command.CrewCreateCommand;
-import org.springframework.web.multipart.MultipartFile;
-
-public interface CrewCommandUseCase {
-    Crew create(CrewCreateCommand command, MultipartFile image);
-}
