--- src/main/java/org/codeit/roomunion/crew/application/port/in/CrewQueryUseCase.java
@@ -1,8 +1,0 @@
-package org.codeit.roomunion.crew.application.port.in;
-
-import org.codeit.roomunion.crew.domain.model.Crew;
-
-public interface CrewQueryUseCase {
-
-    Crew getByCrewId(Long crewId);
-}
