--- src/main/java/org/codeit/roomunion/crew/domain/model/command/CrewCreateCommand.java
@@ -1,33 +1,0 @@
-package org.codeit.roomunion.crew.domain.model.command;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import org.codeit.roomunion.crew.domain.model.enums.CrewCategory;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-@Getter
-@Builder
-@AllArgsConstructor
-public class CrewCreateCommand {
-
-    private final String name;
-
-    private final String description;
-
-    private final CrewCategory category;
-
-    private final int maxMemberCount;
-
-    private final List<String> platformURL;
-
-    private final Long userId;
-
-    private final String imageUrl;
-
-    private final LocalDateTime createdAt;
-
-
-}
