--- src/main/java/org/codeit/roomunion/crew/domain/model/enums/CrewCategory.java
@@ -1,6 +1,0 @@
-package org.codeit.roomunion.crew.domain.model.enums;
-
-public enum CrewCategory {
-    CULTURE_ART, GAME, SELF_DEVELOPMENT, COMMUNICATION, INFO_ECONOMY, HOBBY,
-    ;
-}
