+++ src/main/java/org/codeit/roomunion/moim/domain/model/enums/CrewCategory.java
@@ -1,0 +1,6 @@
+package org.codeit.roomunion.moim.domain.model.enums;
+
+public enum CrewCategory {
+    CULTURE_ART, GAME, SELF_DEVELOPMENT, COMMUNICATION, INFO_ECONOMY, HOBBY,
+    ;
+}
