+++ src/main/java/org/codeit/roomunion/user/domain/command/UserCreateCommand.java
@@ -1,7 +1,7 @@
 package org.codeit.roomunion.user.domain.command;
 
 import lombok.Getter;
-import org.codeit.roomunion.crew.domain.model.enums.CrewCategory;
+import org.codeit.roomunion.moim.domain.model.enums.CrewCategory;
 import org.codeit.roomunion.user.domain.model.Gender;
 
 import java.util.Set;
